[#246]test: 기존 User 테스트 도메인 안에 있던 CouponQueryServiceTest Coupon 테스트 도메인으로 이동
